backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.1
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
dj-database-url==3.1.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
filelock==3.20.2
fsspec==2025.12.0
gguf==0.17.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.4
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mlx==0.30.1
mlx-lm==0.30.2
mlx-metal==0.30.1
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.4.0
openai==2.14.0
packaging==25.0
pandas==2.3.3
pgvector==0.4.2
postgrest==2.27.0
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pymongo==4.15.5
pyparsing==3.3.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==5.0.0rc1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
yarl==1.22.0
